Updated copy function and adjusted operator "=" for dynamic array, updated output and input wallet to file, fixed memory problems of functions and completed income sources and expense categories

+++ Project/Data Structures/Wallet Info/Wallet.h
@@ -14,7 +14,7 @@ struct Wallet {
     func<IncomeSource> is;
     func<ExpenseCategory> ec;
 
-    int totalBalance();
+    long long totalBalance();
 
     Wallet();
     ~Wallet();
